refactor: add missing ng prefix to the angular module name
replace module name from 'ovh-angular-sidebar-menu' to 'ngOvhSidebarMenu'

BREAKING CHANGE: module is now named as `ngOvhSidebarMenu`
