+++ CONTRIBUTING.md
@@ -1,17 +1,15 @@
-# Contributing to ovh-angular-sidebar-menu
+# Contributing to ng-ovh-sidebar-menu
 
 This project accepts contributions. In order to contribute, you should
 pay attention to a few things:
 
 1. your code must follow the coding style rules
-2. your code must be unit-tested
-3. your code must be documented
-4. your work must be signed (see below)
-5. you may contribute through GitHub Pull Requests
+2. your code must be documented
+3. you may contribute through GitHub Pull Requests
 
 # Coding and documentation Style
 
-Refer to [https://github.com/ovh-ux/eslint-config-ovh](https://github.com/ovh-ux/eslint-config-ovh)
+Refer to [eslint-config-airbnb-base](https://github.com/airbnb/javascript/tree/master/packages/eslint-config-airbnb-base)
 
 # Submitting Modifications
 
@@ -19,8 +17,8 @@ The contributions should be submitted through Github Pull Requests.
 
 # Licensing for new files
 
-ovh-angular-sidebar-menu is licensed under a BSD-3-Clause license. Anything
-contributed to ovh-angular-sidebar-menu must be released under this license.
+ng-ovh-sidebar-menu is licensed under a BSD-3-Clause license. Anything
+contributed to ng-ovh-sidebar-menu must be released under this license.
 
 When introducing a new file into the project, please make sure it has a
 copyright header making clear under which license it's being released.
