+++ package.json
@@ -1,19 +1,19 @@
 {
-  "name": "ovh-angular-sidebar-menu",
+  "name": "@ovh-ux/ng-ovh-sidebar-menu",
   "version": "8.0.0-alpha.4",
   "description": "Manage and display a left menu tree",
   "keywords": [
     "angular"
   ],
-  "repository": "ovh-ux/ovh-angular-sidebar-menu",
+  "repository": "ovh-ux/ng-ovh-sidebar-menu",
   "license": "BSD-3-Clause",
   "author": "OVH SAS",
   "files": [
     "dist"
   ],
-  "main": "./dist/cjs/ovh-angular-sidebar-menu.js",
-  "module": "./dist/esm/ovh-angular-sidebar-menu.js",
-  "browser": "./dist/umd/ovh-angular-sidebar-menu.js",
+  "main": "./dist/cjs/index.js",
+  "module": "./dist/esm/index.js",
+  "browser": "./dist/umd/ng-ovh-sidebar-menu.js",
   "scripts": {
     "build": "rollup -c",
     "dev": "rollup -c --watch",
